+++ src/controllers/MessageController.php
@@ -45,6 +45,16 @@ class MessageController extends BaseController
     /**
      * @var string
      */
+    const EVENT_BEFORE_VALIDATE = "beforeValidate";
+
+    /**
+     * @var string
+     */
+    const EVENT_AFTER_VALIDATE = "afterValidate";
+
+    /**
+     * @var string
+     */
     const EVENT_BEFORE_RESPONSE = "beforeResponse";
 
     /**
@@ -111,6 +121,16 @@ class MessageController extends BaseController
             }
         }
 
+        $beforeValidateEvent = new MessageEvent([
+            'form_id' => $this->formModel->id,
+            'message' => $request->post(),
+            'errors' => $errors,
+        ]);
+
+        //Trigger Event to allow plugins to do custom validation to the values
+        $this->trigger(self::EVENT_BEFORE_VALIDATE, $beforeValidateEvent);
+        $errors = $beforeValidateEvent->errors;
+
         $visualFields = FormField::getVisualFields();
         if(empty($errors)) {
             // Get Form Fields to validate them
@@ -131,6 +151,16 @@ class MessageController extends BaseController
             }
         }
 
+        //Trigger Event to allow plugins to do custom validation to the values
+        $event = new MessageEvent([
+            'form_id' => $this->formModel->id,
+            'message' => $entryValues,
+            'errors' => $errors,
+        ]);
+        $this->trigger(self::EVENT_AFTER_VALIDATE, $event);
+
+        $errors = $event->errors;
+
         if (! empty($errors)) {
             $response = [
                 'values' => $request->post(),
@@ -150,10 +180,6 @@ class MessageController extends BaseController
         }
 
         //Trigger Event to allow plugins to modify fields before being saved to the database
-        $event = new MessageEvent([
-            'form_id' => $this->formModel->id,
-            'message' => $entryValues
-        ]);
         $this->trigger(self::EVENT_BEFORE_SAVE, $event);
 
         // Values for Mailer
@@ -234,6 +260,7 @@ class MessageController extends BaseController
 
         Craft::$app->getSession()->setNotice($responseEvent->message);
         Craft::$app->getSession()->setFlash('wheelformSuccess',$responseEvent->message);
+        Craft::$app->getSession()->setFlash('wheelformSubmittedForm', $this->formModel->id);
 
         return $this->redirectToEventUrl($responseEvent->redirect, $message);
     }