+++ src/services/FormService.php
@@ -2,6 +2,7 @@
 namespace wheelform\services;
 
 use Craft;
+use craft\helpers\App;
 use craft\helpers\Template;
 use Exception;
 use wheelform\db\Form;
@@ -114,7 +115,7 @@ class FormService extends BaseService
     {
         $html = '';
         $settings = Wheelform::getInstance()->getSettings();
-        $recaptcha_public = empty($settings['recaptcha_public']) ? "" : Craft::parseEnv($settings['recaptcha_public']);
+        $recaptcha_public = empty($settings['recaptcha_public']) ? "" : App::parseEnv($settings['recaptcha_public']);
         if(intval($this->instance->recaptcha) && ! empty($recaptcha_public)) {
             if(! empty($settings['recaptcha_version'] && $settings['recaptcha_version'] == '3')) {
                 $html .= $this->renderRecaptchaV3Event();