+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## 2.6.5 - 2022-03-04
+### Added
+- Telephone input field type
+- Date input field type
+- Time input field type
+- New beforeValidate and afterValidate events for plugins to do custom validation
+- New flash variable with submitted form ID
+### Fix
+- Duplicate fields in CSV Export when fields are empty.
+- Logic with multiple fields of same type to have increasing names to disallow having the same input field name.
+
 ## 2.6.4 - 2022-01-19
 ### Added
 - Ability to modify redirect using events
 - New multiple option for select fields
-
 
 ## 2.6.3 - 2021-11-28
 ### Added